Rename Renamed a few files and made associated changes main.c -> main_pru1.c resource_table_empty.h -> resource_table_pru1.h
