+++ examples/firmware_exmples/PRU_inline_asm_blinky/main_pru1.c
@@ -36,7 +36,7 @@
 				  
 #include <stdint.h>
 #include <pru_cfg.h>
-#include "resource_table_empty.h"
+#include "resource_table_pru1.h"
 
 // The function is defined in pru1_asm_blinky.asm in same dir
 // We just need to add a declaration here, the defination can be
