+++ examples/firmware_exmples/PRU_inline_asm_blinky/resource_table_pru1.h
